libstdc++: PSTL dispatch for C++20 range random access iterators
This patch closes 110512, in which random access iterators
originated from C++20 ranges are dispatched to the sequential
algorithms in the parallel STL because, before this patch,
the parallel STL checks whether an iterator inherits from the
random access iterator tag. This patch extends the check
in C++ >= 20 to also check whether the iterator models the
std::random_access_iterator concept. This is allowed
by C++23's P2408, which is safe to backport to C++20 because
any application that would break already exhibits undefined
behavior due to pre-condition violation.